Document threadsafety of isready(::Channel) and remove data race
Make it clear that isready() is not threadsafe, as it has a data race.
We could make it threadsafe with lock() but this would also make it
blocking and wouldn't make uses of it non-racy without a lock around the
surronding code. So for now just document the current state of affairs.

Also remove an optimistic use of `isready()` in `wait()` as it's a data
race.